+++ mk.h
@@ -8,7 +8,7 @@
 #include<pthread.h>
 
 typedef struct {
-    int db_size;
+    size_t db_size;
     char *db_mk_as1;
     char *db_mk_as2;
     unsigned char *db_mk_ba;
@@ -18,9 +18,9 @@ typedef struct {
 } db_mk_infomation;
 
 
-int db_mk_init(void *mk_info, int db_size);
-void *mk_read(int index);
-int mk_write(int index, void* value);
+int db_mk_init(void *mk_info, size_t db_size);
+void *mk_read(size_t index);
+int mk_write(size_t index, void* value);
 void db_mk_ckp(int ckp_order, void *mk_info);
 void db_mk_destroy(void *mk_info);
